refactor: separate Tokenizer from StringCalculator
