+++ src/main/java/calculator/domain/Tokenizer.java
@@ -1,6 +1,5 @@
-package calculator.util;
+package calculator.domain;
 
-import calculator.domain.NumberTokens;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -19,6 +18,18 @@ public class Tokenizer {
     public Tokenizer(String text) {
         this.numberTokens = new NumberTokens(tokenize(text));
         System.out.println();
+    }
+
+    public static List<Integer> parseInt(String[] tokens) {
+        try {
+            return Arrays.stream(tokens).map(Integer::parseInt).collect(Collectors.toList());
+        } catch (NumberFormatException | NullPointerException e) {
+            throw new IllegalArgumentException("[ERROR] 지정된 구분자 외의 문자열이 있습니다.");
+        }
+    }
+
+    public static boolean isCustomDelimiter(String text) {
+        return Pattern.compile("//(.)\n(.*)").matcher(text).find();
     }
 
     private List<Integer> tokenize(String text) {
@@ -46,18 +57,6 @@ public class Tokenizer {
 
     private boolean isEmptyOrNull(String text) {
         return Objects.isNull(text) || text.isEmpty();
-    }
-
-    public static List<Integer> parseInt(String[] tokens) {
-        try {
-            return Arrays.stream(tokens).map(Integer::parseInt).collect(Collectors.toList());
-        } catch (NumberFormatException | NullPointerException e) {
-            throw new IllegalArgumentException("[ERROR] 지정된 구분자 외의 문자열이 있습니다.");
-        }
-    }
-
-    public static boolean isCustomDelimiter(String text) {
-        return Pattern.compile("//(.)\n(.*)").matcher(text).find();
     }
 
     public NumberTokens getTokenizedNumbers() {
